Show that you can import regular packages from the namespace packages
