+++ allennlp/tests/fixtures/plugins/project_a/allennlp_plugins/a/__init__.py
@@ -1,10 +1,1 @@
-import argparse
-
-from allennlp.commands import Subcommand
-
-
-class A(Subcommand):
-    def add_subparser(
-        self, name: str, parser: argparse._SubParsersAction
-    ) -> argparse.ArgumentParser:
-        raise NotImplementedError
+from a.a import A
