+++ allennlp/tests/fixtures/plugins/project_b/allennlp_plugins/b/__init__.py
@@ -1,10 +1,1 @@
-import argparse
-
-from allennlp.commands import Subcommand
-
-
-class B(Subcommand):
-    def add_subparser(
-        self, name: str, parser: argparse._SubParsersAction
-    ) -> argparse.ArgumentParser:
-        raise NotImplementedError
+from b.b import B
